# public/swagger.yaml
openapi: 3.0.0
info:
  title: 'Laravel App'
  version: '1.0'
servers:
  -
    url: 'http://laravel-app.local'
paths:
  /api/account/show:
    get:
      tags:
        - account
      operationId: 7b10175533fd9d5dd4be1d8b950e90ec
      responses:
        '200':
          description: 'Account data'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/account/register:
    post:
      tags:
        - account
      operationId: dd1184c14a98467480820df6a52e04bb
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  nullable: false
                email:
                  type: string
                  nullable: false
                password:
                  type: string
                  format: password
                  nullable: false
              type: object
      responses:
        '201':
          description: 'Account data'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/account/update:
    put:
      tags:
        - account
      operationId: 5cfed2aa6285b24797f79bc20c3f7106
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  nullable: true
                email:
                  type: string
                  nullable: true
                username:
                  type: string
                  nullable: true
                password:
                  type: string
                  format: password
                  nullable: true
                discord_user_id:
                  type: string
                  nullable: true
              type: object
      responses:
        '200':
          description: 'Account data'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/account/settings:
    put:
      tags:
        - account
      operationId: 2eacf149a4144a183a782362f4140e28
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                platforms:
                  type: array
                  items: { type: string, enum: [pc, playstation5, xbox-series-x, switch, linux, android, ios] }
                  nullable: true
                genres:
                  type: array
                  items: { type: string, enum: [racing, shooter, adventure, action, rpg, fighting, puzzle, strategy, arcade, simulation, sports, card, family, board-games, educational, casual, indie, massively-multiplayer, platformer] }
                  nullable: true
                period:
                  type: string
                  enum: [next-7-days, next-30-days, next-12-months]
                  nullable: true
                frequency:
                  type: string
                  enum: [none, daily, weekly, monthly]
                  nullable: true
              type: object
      responses:
        '200':
          description: 'Account data'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/auth/login:
    post:
      tags:
        - auth
      operationId: 8dcb70df1020986038d098cc08d05dae
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
              type: object
      responses:
        '200':
          description: JWT
      security: []
  /api/rawg/domain/genres:
    get:
      tags:
        - domain
      operationId: daa3484b0cfbf1f84c09190d386fb6aa
      responses:
        '200':
          description: 'List of RAWG genres'
  /api/rawg/domain/tags:
    get:
      tags:
        - domain
      operationId: b51af1f826392611185f7b288fedb845
      responses:
        '200':
          description: 'List of RAWG tags'
  /api/rawg/domain/platforms:
    get:
      tags:
        - domain
      operationId: b1a9bad8e0b69f95fd054ffcf6c4b9b3
      responses:
        '200':
          description: 'List of RAWG platforms'
  '/api/rawg/games/recommendations/{genre}':
    get:
      tags:
        - games
      operationId: 9ce6b8a31793e7f454284c21b860704b
      parameters:
        -
          name: genre
          in: path
          description: 'Filter by genres'
          required: true
          schema:
            $ref: '#/components/schemas/RawgGenre'
        -
          name: platforms
          in: query
          description: 'Filter by platforms (accepts comma separated list)'
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
              enum:
                - pc
                - playstation5
                - xbox-series-x
                - switch
                - linux
                - android
                - ios
        -
          name: ordering
          in: query
          description: 'Rawg field to order by'
        -
          name: page
          in: query
          description: 'Page number to request'
        -
          name: page_size
          in: query
          description: 'How many items per page'
      responses:
        '200':
          description: 'List of RAWG games'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  '/api/rawg/games/upcoming-releases/{period}':
    get:
      tags:
        - games
      operationId: afce02b91592c98f28a021c4a88bbfde
      parameters:
        -
          name: period
          in: path
          description: 'Get releases for selected period'
          required: true
          schema:
            $ref: '#/components/schemas/Period'
        -
          name: genres
          in: query
          description: 'Filter by genres (accepts comma separated list)'
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
              enum:
                - racing
                - shooter
                - adventure
                - action
                - rpg
                - fighting
                - puzzle
                - strategy
                - arcade
                - simulation
                - sports
                - card
                - family
                - board-games
                - educational
                - casual
                - indie
                - massively-multiplayer
                - platformer
        -
          name: platforms
          in: query
          description: 'Filter by platforms (accepts comma separated list)'
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
              enum:
                - pc
                - playstation5
                - xbox-series-x
                - switch
                - linux
                - android
                - ios
        -
          name: ordering
          in: query
          description: 'Rawg field to order by'
        -
          name: page
          in: query
          description: 'Page number to request'
        -
          name: page_size
          in: query
          description: 'How many items per page'
      responses:
        '200':
          description: 'List of RAWG games'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  '/api/rawg/games/{game}/achievements':
    get:
      tags:
        - games
      operationId: 6eca34be51c7b0acaf24caf35cbd7a90
      parameters:
        -
          name: game
          in: path
          description: 'Rawg slug of the game'
          required: true
        -
          name: order_by
          in: query
          description: 'Field to order by'
        -
          name: sort_order
          in: query
          description: 'Sorting order'
          schema:
            $ref: '#/components/schemas/SortOrder'
        -
          name: page
          in: query
          description: 'Page number to request'
        -
          name: page_size
          in: query
          description: 'How many items per page'
      responses:
        '200':
          description: "List of RAWG game's achievements"
          content:
            application/json:
              schema:
                type: array
                items:
                  properties: { id: { type: integer }, name: { type: string }, description: { type: string }, image: { type: string }, percent: { type: string } }
                  type: object
  /up:
    get:
      tags:
        - application
      operationId: babf1c1e7a1e559b1836be5e2e457254
      responses:
        '200':
          description: OK
      security: []
  /api/health:
    get:
      tags:
        - application
      operationId: cf1f080fb1aecf4993751251dedf0b57
      responses:
        '200':
          description: OK
      security: []
components:
  schemas:
    Acknowledge:
      type: integer
      enum:
        - Pong
        - ChannelMessageWithSource
        - UpdateMessage
        - AutoCompleteResult
    ButtonStyle:
      type: integer
      enum:
        - Primary
        - Secundary
    ComponentType:
      type: integer
      enum:
        - ActionRow
        - Button
        - StringSelect
    InteractionType:
      type: integer
      enum:
        - Ping
        - Command
        - MessageComponent
        - AutoComplete
    OptionType:
      type: integer
      enum:
        - String
    Frequency:
      type: string
      enum:
        - none
        - daily
        - weekly
        - monthly
    Period:
      type: string
      enum:
        - next-7-days
        - next-30-days
        - next-12-months
    Platform:
      type: string
      enum:
        - pc
        - playstation5
        - xbox-series-x
        - switch
        - linux
        - android
        - ios
    RawgField:
      type: string
      enum:
        - id
        - name
        - slug
        - background_image
        - released
        - dates
        - genres
        - platforms
        - stores
        - ordering
        - page_size
        - page
    RawgGenre:
      type: string
      enum:
        - racing
        - shooter
        - adventure
        - action
        - rpg
        - fighting
        - puzzle
        - strategy
        - arcade
        - simulation
        - sports
        - card
        - family
        - board-games
        - educational
        - casual
        - indie
        - massively-multiplayer
        - platformer
    RawgPlatform:
      type: integer
      enum:
        - 4
        - 187
        - 186
        - 7
        - 6
        - 21
        - 3
    Scope:
      type: string
      enum:
        - default
        - root
    SortOrder:
      type: string
      enum:
        - ASC
        - DESC
    Game:
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        background_image:
          type: string
        released:
          type: string
          format: date-time
        platforms:
          type: array
          items:
            properties:
              id:
                type: integer
              name:
                type: string
              slug:
                type: string
            type: object
        stores:
          type: array
          items:
            type: object
        genres:
          type: array
          items:
            type: object
      type: object
    PaginatedResponse:
      properties:
        total:
          type: integer
        pageSize:
          type: integer
        currentPage:
          type: integer
        lastPage:
          type: integer
        nextPageUrl:
          type: string
        prevPageUrl:
          type: string
        data:
          type: array
          items:
            oneOf:
              -
                $ref: '#/components/schemas/Game'
      type: object
    User:
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        discord_user_id:
          type: string
        discord_username:
          type: string
        discord_channel_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        settings:
          $ref: '#/components/schemas/UserSetting'
      type: object
    UserSetting:
      properties:
        id:
          type: integer
        user_id:
          type: integer
        platforms:
          type: array
          items:
            type: string
            enum:
              - pc
              - playstation5
              - xbox-series-x
              - switch
              - linux
              - android
              - ios
        genres:
          type: array
          items:
            type: string
            enum:
              - racing
              - shooter
              - adventure
              - action
              - rpg
              - fighting
              - puzzle
              - strategy
              - arcade
              - simulation
              - sports
              - card
              - family
              - board-games
              - educational
              - casual
              - indie
              - massively-multiplayer
              - platformer
        period:
          type: string
          enum:
            - next-7-days
            - next-30-days
            - next-12-months
        frequency:
          type: string
          enum:
            - none
            - daily
            - weekly
            - monthly
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      in: header
      scheme: bearer
security:
  -
    bearerAuth: []
tags:
  -
    name: application
    description: 'health and other application routes'
  -
    name: domain
    description: 'rawg domain routes'
  -
    name: games
    description: 'rawg games routes'
  -
    name: auth
    description: 'authentication routes'
  -
    name: account
    description: 'account management routes'